+++ changelog.md
@@ -7,6 +7,7 @@
 
 - Add support for `only_platforms` and `skip_platforms` settings in test plans, to specify platforms to only runs tests on and/or to explicitly skips.
 - Add a `--skip-platform`/`-s` option to `gha-utils test-plan` to allow skipping of test plan on a whole set of platforms.
+- Allow `--plan` option to be used multiple times in `gha-utils test-plan` to merge multiple test plans.
 - Use `windows-2025` instead of `windows-2022` for Nuitka builds.
 - Add `windows-2025` to the test matrix.
 - Fix running of Nuitka-compiled `gha-utils metadata` command.